requirements.txt
Flask
Flask-Cors
ffmpeg-python